+++ base/file.jl
@@ -604,7 +604,7 @@ function mktempdir(parent::AbstractString=tempdir();
 
     req = Libc.malloc(_sizeof_uv_fs)
     try
-        ret = ccall(:uv_fs_mkdtemp, Int32,
+        ret = ccall(:uv_fs_mkdtemp, Cint,
                     (Ptr{Cvoid}, Ptr{Cvoid}, Cstring, Ptr{Cvoid}),
                     C_NULL, req, tpath, C_NULL)
         if ret < 0
@@ -737,7 +737,7 @@ julia> readdir(abspath("base"), join=true)
  "/home/JuliaUser/dev/julia/base/weakkeydict.jl"
 ```
 """
-function readdir(dir::AbstractString=pwd(); join::Bool=false)
+function readdir(dir::AbstractString; join::Bool=false)
     # Allocate space for uv_fs_t struct
     uv_readdir_req = zeros(UInt8, ccall(:jl_sizeof_uv_fs_t, Int32, ()))
 
@@ -759,6 +759,7 @@ function readdir(dir::AbstractString=pwd(); join::Bool=false)
 
     return entries
 end
+readdir(; join::Bool=false) = readdir(join ? pwd() : ".", join=join)
 
 """
     walkdir(dir; topdown=true, follow_symlinks=false, onerror=throw)