+++ base/locks-mt.jl
@@ -30,6 +30,10 @@ struct SpinLock <: AbstractLock
     SpinLock() = new(Atomic{Int}(0))
 end
 
+# Note: this cannot assert that the lock is held by the correct thread, because we do not
+# track which thread locked it. Users beware.
+Base.assert_havelock(l::SpinLock) = islocked(l) ? nothing : concurrency_violation()
+
 function lock(l::SpinLock)
     while true
         if l.handle[] == 0