+++ base/ryu/shortest.jl
@@ -1,6 +1,6 @@
 @inline function writeshortest(buf::Vector{UInt8}, pos, x::T,
     plus=false, space=false, hash=true,
-    precision=-1, expchar=UInt8('e'), padexp=false, decchar=UInt8('.')) where {T}
+    precision=-1, expchar=UInt8('e'), padexp=false, decchar=UInt8('.'), typed=false) where {T}
     @assert 0 < pos <= length(buf)
     neg = signbit(x)
     # special cases
@@ -23,19 +23,39 @@
         end
         if precision == -1
             buf[pos] = UInt8('0')
-            return pos + 1
+            pos += 1
+            if typed && x isa Float32
+                buf[pos] = UInt8('f')
+                buf[pos + 1] = UInt8('0')
+                pos += 2
+            end
+            return pos
         end
         while precision > 1
             buf[pos] = UInt8('0')
             pos += 1
             precision -= 1
         end
+        if typed && x isa Float32
+            buf[pos] = UInt8('f')
+            buf[pos + 1] = UInt8('0')
+            pos += 2
+        end
         return pos
     elseif isnan(x)
         buf[pos] = UInt8('N')
         buf[pos + 1] = UInt8('a')
         buf[pos + 2] = UInt8('N')
-        return pos + 3
+        if typed
+            if x isa Float32
+                buf[pos + 3] = UInt8('3')
+                buf[pos + 4] = UInt8('2')
+            elseif x isa Float16
+                buf[pos + 3] = UInt8('1')
+                buf[pos + 4] = UInt8('6')
+            end
+        end
+        return pos + 3 + (typed && x isa Union{Float32, Float16} ? 2 : 0)
     elseif !isfinite(x)
         if neg
             buf[pos] = UInt8('-')
@@ -43,7 +63,16 @@
         buf[pos + neg] = UInt8('I')
         buf[pos + neg + 1] = UInt8('n')
         buf[pos + neg + 2] = UInt8('f')
-        return pos + neg + 3
+        if typed
+            if x isa Float32
+                buf[pos + neg + 3] = UInt8('3')
+                buf[pos + neg + 4] = UInt8('2')
+            elseif x isa Float16
+                buf[pos + neg + 3] = UInt8('1')
+                buf[pos + neg + 4] = UInt8('6')
+            end
+        end
+        return pos + neg + 3 + (typed && x isa Union{Float32, Float16} ? 2 : 0)
     end
 
     bits = uint(x)
@@ -191,6 +220,17 @@
         end
     end
 
+    if typed && x isa Float16
+        buf[pos] = UInt8('F')
+        buf[pos + 1] = UInt8('l')
+        buf[pos + 2] = UInt8('o')
+        buf[pos + 3] = UInt8('a')
+        buf[pos + 4] = UInt8('t')
+        buf[pos + 5] = UInt8('1')
+        buf[pos + 6] = UInt8('6')
+        buf[pos + 7] = UInt8('(')
+        pos += 8
+    end
     if neg
         buf[pos] = UInt8('-')
         pos += 1
@@ -205,7 +245,7 @@
     olength = decimallength(output)
     exp_form = true
     pt = nexp + olength
-    if -4 < pt <= (precision == -1 ? (T == Float16 ? 5 : 6) : precision) #&& !(pt >= olength && abs(mod(x + 0.05, 10^(pt - olength)) - 0.05) > 0.05)
+    if -4 < pt <= (precision == -1 ? (T == Float16 ? 5 : 6) : precision)
         exp_form = false
         if pt <= 0
             buf[pos] = UInt8('0')
@@ -311,6 +351,11 @@
                 precision -= 1
             end
         end
+        if typed && x isa Float32
+            buf[pos] = UInt8('f')
+            buf[pos + 1] = UInt8('0')
+            pos += 2
+        end
     else
         if olength > 1 || hash
             buf[pos] = decchar
@@ -355,6 +400,10 @@
             buf[pos] = UInt8('0') + (exp2 % UInt8)
             pos += 1
         end
+    end
+    if typed && x isa Float16
+        buf[pos] = UInt8(')')
+        pos += 1
     end
 
     return pos