+++ test/file.jl
@@ -1340,7 +1340,7 @@ mktempdir() do dir
             @test basename(realpath(uppercase(path))) == path
         end
         rm(path)
-        @test_throws SystemError realpath(path)
+        @test_throws Base.IOError realpath(path)
     end
 end
 
@@ -1405,7 +1405,7 @@ end
     end
 end
 
-@testset "readir tests" begin
+@testset "readdir tests" begin
     ≛(a, b) = sort(a) == sort(b)
     mktempdir() do dir
         d = cd(pwd, dir) # might resolve symlinks
@@ -1429,6 +1429,23 @@ end
         cd(t) do
             @test readdir(b) ≛ names
             @test readdir(b, join=true) ≛ [joinpath(b, x) for x in names]
+        end
+    end
+    if !Sys.iswindows()
+        mktempdir() do dir
+            cd(dir) do
+                d = pwd() # might resolve symlinks
+                @test isdir(d)
+                @test Base.Filesystem.samefile(d, ".")
+                @test isempty(readdir())
+                @test isempty(readdir(d))
+                @test isempty(readdir(join=true))
+                rm(d, recursive=true)
+                @test !ispath(d)
+                @test isempty(readdir())
+                @test_throws SystemError readdir(d)
+                @test_throws Base.IOError readdir(join=true)
+            end
         end
     end
 end