+++ NEWS.md
@@ -6,9 +6,12 @@ New language features
 
 * Structs with all isbits and isbitsunion fields are now stored inline in arrays ([#32448]).
 
+* `import` now allows quoted symbols, e.g. `import Base.:+` ([#33158]).
+
 Language changes
 ----------------
 
+* Calling `show` or `repr` on an `undef`/`UndefInitializer()` array initializer now shows valid Julia code ([#33211]).
 
 Multi-threading changes
 -----------------------
@@ -27,6 +30,7 @@ New library functions
 * The `readdir` function now takes a `join::Bool` keyword argument defaulting to `false`, which when set causes `readdir` to join its directory argument with each listed name ([#33113]).
 * The new `only(x)` function returns the one-and-only element of a collection `x`, and throws an `ArgumentError` if `x` contains zero or multiple elements. ([#33129])
 
+
 Standard library changes
 ------------------------
 
@@ -38,6 +42,7 @@ Standard library changes
 
 * `qr` and `qr!` functions support `blocksize` keyword argument ([#33053]).
 
+* `dot` now admits a 3-argument method `dot(x, A, y)` to compute generalized dot products `dot(x, A*y)`, but without computing and storing the intermediate result `A*y` ([#32739]).
 
 #### SparseArrays